refactor-#158: move booking controller to controllers
