+++ src/microservices/booking/rest/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from './booking.controller';
